+++ config/src/main/resources/profiles/dev.toml
@@ -1,0 +1,1 @@
+network="DEV"
