--- config/src/main/resources/profiles/staker.toml
@@ -1,1 +1,0 @@
-sync-mode="FULL"
